Refactor NCCL instance launch functions and increase wait time and count
Cleanup the NCCL instance launch functions so that they are a bit more
readable. There were too many loops to follow, gotos, and reliance on
global variables for this to make sense. Plus, the inner loop in
prepare_instance wasn't retrying anyway.

Because the NCCL test jobs run for 15+ minutes, increase the number of
retries and the amount of time to wait between trying to launch in all
AZs in a region.